Implement custom usage of berries
Implement options to customize the usage of berries.
Berries will now be used basing on the probability of the catch and then
either UseBerryMinCp or UseBerryMinIv will be evaluated based if
PrioritizeIvOverCp is set to true or false.
Additionally, in case of berry usage a correct string will be printed.