Teleport AI (Need Help)
This new Teleport AI will Learn as it goes. It still needs some more
work but I'm almost ready for release. It generates a new file called
TeleAI.json in the config file. It can load the file but im having
trouble saving the new delays. As soon as that is done it is ready for
release.